refactor(assistant-stream-py): extract heartbeat machinery into serialization/heartbeat.py and retrieve the cancelled task outcome
